requirements.txt
flask
tensorflow 
keras 
numpy 
pillow 
google-generativeai 
python-dotenv